+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - Support for running tests on [CircleCI](https://circleci.com/).
 
+### Fixed
+- Return proper descriptions for failing tabular facts.
+
 ## [1.0.0] - 2018-11-12
 
 ### Added